+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Added
 
+- Added support for matcher-combinators
+
 ## Fixed
 
 ## Changed
+
+- When using the nodejs repl type, automatically set the CLJS compile target to :nodejs
 
 # 0.0-51 (2019-09-11 / 4b6a751)